refactor: Add .js extensions
