+++ src/index.ts
@@ -1,8 +1,8 @@
-export { default as LocalAPI } from './local-api';
-export { default as Message } from './message';
-export { default as MultipleBoards } from './multiple-boards';
-export { default as ReadWriteAPI } from './read-write-api';
-export { default as SubscriptionAPI } from './subscription-api';
+export { default as LocalAPI } from './local-api.js';
+export { default as Message } from './message.js';
+export { default as MultipleBoards } from './multiple-boards.js';
+export { default as ReadWriteAPI } from './read-write-api.js';
+export { default as SubscriptionAPI } from './subscription-api.js';
 export {
     BOARD_LINE_LENGTH,
     BOARD_LINES,
