+++ src/message.ts
@@ -1,5 +1,5 @@
 import GraphemeSplitter from 'grapheme-splitter';
-import { BOARD_LINE_LENGTH, BOARD_LINES, BoardCharArray, MessageWriteOptions, MessageWritePosition } from './types';
+import { BOARD_LINE_LENGTH, BOARD_LINES, BoardCharArray, MessageWriteOptions, MessageWritePosition } from './types.js';
 
 
 /**
